pyproject.toml
[project]
name = "langgraph-projects"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.30.1",
    "langchain-openai>=0.3.32",
    "langchain[anthropic]>=0.3.27",
    "langgraph>=0.6.6",
    "python-dotenv>=1.1.1",
]
